last state before merge
